Add image alignment feature
This allows specifying an (optional) alignment in an {img} tag
replacement, so the full form now is {img:src:width:height[:align]}.

If the alignment is not given, or is otherwise invalid, it will
default to using center to match existing behaviour.

Unfortunately it seems the WoW implementation of this has some
issues however; images aligned either left or right don't seem to
stack correctly and not overlap in the same way centered ones do.

Also relocated some upvalue declarations to the header, and removed
a duplicate strtrim capture.