Disable map scan button until broadcast is ready
This adds a few lifecycle events for the broadcast system which
are used by the map module to control the button state.

The initial state of the button will be enabled and clickable,
but when BROADCAST_CHANNEL_CONNECTING is fired we'll disable
the button and update the tooltip accordingly.

If an error occurs in setup it should display in the tooltip
and prevent the button from activating, but this isn't yet tested.

Also not yet checked what happens if broadcast is disabled; my
guess is the button will stay enabled.