Implement reordering of additional characteristics
The list can be reordered while editing the profile by dragging and
dropping the icon (we're using that as the handle). The values
correctly persist when saved.

The drag/drop operation is immediate, so when you move an item over
another it'll take its place rather than waiting for you to release
the mouse button. This is a bit simpler to implement, since otherwise
we'd need an indicator texture to say where it's going to drop.

In addition, if the additional info fields are near the boundaries
of the scroll frame, it'll scroll in the correct direction automatically
if you go past them so you can see where you're dragging to.