Fix scroll adjustment logic
Move the responsibility for checking if the mouse isn't over the
frame to miscInfoScrollParentTowardCursor().

If the cursor isn't between the x1 and x2 coordinates, don't apply
any adjustments to the scroll. This prevents issues where the frame
would scroll for no good reason if you were dragging items off to the
side.

Also somewhat clarify the return value name from
miscInfoTestRectangleCoord. "relIndex" isn't truthful since it's not an
index, really.