Profile migration for upscaled psycho values
Adds a V2 field to the psycho characteristics structure which
stores an upscaled equivalent of the VA field, ranging from 0 to 20.

The UI hasn't been implemented yet, however the necessary code is in
place in the characteristics display page to correctly use the V2 field
instead of the VA.

Some constants have been added to replace the hardwired numbers around
the place.

In addition, updates to the V2 field will propagate backwards to the VA
field; this means that older clients will still be able to view changes
made to profiles by users running a newer version.