First pass at UI implementation
Use a statusbar based on the ones used in the Achievement UI to display
the value. In edit mode, we overlay a slider on top of it that only has
a thumb texture.

Hook it all up and made the UI functional, so you can see the values and
edit them. Lacking the stack count on the icons, and the colours are
temporary right now.

Added parentKeys to all the elements of the PsychoInfoLineBar (and edit
equivalent) templates and swapped to those in the main code, since the
name concatenations weren't particularly easy to read. The names are
still present in case any third parties seem to rely on
them, but the statusbar itself is named "$parentBar" rather than
"$parentJauge" - this seems acceptable to me at least since we're
entirely changing out the frame type.

The existing template for the dot-based display exists, again just in
case there's a third party using it for some reason.