Add basic psycho trait bar colouring
The implementation doesn't include a colour picker to customise the
colours of the bars, however there is support for reading the necessary
value from the data structure in storage using the LC/RC fields (which
are treated as hex-encoded colours).

As there's no colour picker and none of the presets have colour
overrides, a global pair of colours have been defined as fallbacks.