Fix upscaling of psycho VA values
Of course one thing I forgot to implement was the case where a user
receives a profile from an older version; VA would be set in that case
but V2 would never get created, so we'd always retrieve the default.

register_characteristics.lua has a getter function
(getPsychoStructureValue) that will now fix that for us however. It'll
use the V2 value if present, if not it'll upscale the VA value and
return that. If there's not a VA value, you'll get the default.

Also no longer apply the migration to the register. There's no point,
since we're handling it via the getter function above now anyway.