Implement dice roll modifiers
Syntax for the roll with a modifier is "/trp3 roll 1d20+10" as
suggested. Supports negative values. The modifier is applied to the
total outcome of the roll (so 2d20+5 would just add +5 to the final
result).

Also fixed up some issues with target validity checking; only send to
players of the same faction and only if they're really players.

Finally, drop the requirement to prefix the roll with the number of
dice - so you can just "/trp3 roll d20" for the equivalent of 1d20.