Swap to using colour swatches for psycho attribs
This does cramp the UI up somewhat but it works better than I'd
expected. Plus it's more consistent, so that's a bonus.

Also cleaned up some code with regards to hiding elements that shouldn't
be shown for custom vs. preset traits; there's a pair of booleans we can
apply to elements (done in the XML right now) and a new function which
runs through a given list of children and toggles their visibility
accordingly.

Cleaned up/reverted any reference to the dropdown code for setting
colours, as well as the localisation string change for that widget.