Implement copy/pasting of glance traits
This is provided as part of the right-click menu. While another ticket
does suggest the use of Ctrl-Click for the copy operation, there's the
question of how do you then paste it - a button could be added to the
editor instead, but this approach feels fine (I mean, /everything/ lets
you right-click for copy and paste normally, right?).

A lot of this would benefit from a better dropdown system, as the
current implementation relies on the glance buttons having global names
to perform the copy at least.