Implementation of map scan blip coloring
This colors blips on the world map according to the relationship status
defined on their profile.

The tooltip display for blips has been improved to split the names into
categories based on the same status; this is ordered logically such that
family members would appear at the top, and the bottom would be profiles
you have no relationship with. Names are sorted in ascending order in each
category.

Swapped to using texture atlas names for marker icons. These map to texcoords
and file IDs managed by Blizzard, so even if things were moved around in
the future (unlikely) they should remain valid.

Moved relationship constants to the Globals table. These are backwards
compatible and are still exported in the same place, but it makes it
easier for other areas to get at them without requiring a profile ID.