Build credits from tables, revamp dashboard
This requires manual changes to the non-English locales that are
outlined below.

Modernised the dashboard whilst aiming to keep it backwards compatible
with any existing assumptions on its structure. The codebase is
consistent with the style being used on the color picker mixin branch,
and is extensible to boot with features like clearly defined
registration points for hyperlink handlers.

TRP3_DashboardTabFrame is a virtual frame with a mixin attached which
matches the old template precisely, same widgets, same sizes, same
positioning. This is responsible for owning the tabs and switching
between them, as well as providing the views access to the concrete
dashboard frame itself. When instantiating this frame, you must supply
an OnLoad handler that defines "self.TabClasses = { classes... }" and
then call "self:OnLoad()".

TRP3_DashboardHTMLContent (and its associated mixin) is a SimpleHTML
frame template which has an overridden SetText() method to provide us
with a GetText(), so the ".text" member is no longer needed. It also
defines the script handlers for hyperlink management as well as a pair
of RegisterHyperlink and UnregisterHyperlink methods. These are used by
the "What's New" tab view to provide a clean API for adding in custom
links for the (amazing) option toggles in the changelog.

DashboardTabView is a base class for implementing the individual tab
views. It's pretty barebones and basic, nothing special.

The individual views subclass DashboardTabView and override Show to
provide logic for updating the HTML content on the dashboard, as well
as other stuff they need. There's also a Hide method which, oddly
enough, is called when the tab is going to be hidden.

The about view has been reworked and the credits defined there with some
helper classes to simplify displaying names, including those special
people who are multi-talented with loads of roles. This includes the use
of Ellyb to get the Patreon contributors list.

Manual localisation changes can be seen in the definition for the
English (default) locale.