Disallow copying of nil glances
The first-glance bar would allow you to copy traits from other peoples
profiles in slots that had no exported trait (such that button.data was
nil). The copy/paste logic was fine with this, but it seemed a bit odd
to let you copy/paste a non-existant trait so we now filter the copy
option and only apply it to buttons that actually have a data field.

In addition, the dropdown on trait buttons will now only show if there's
any actual items to display - so right-clicking another person's
non-existant trait won't leave you with a dropdown that just says
"Cancel".