Refactored/simplified, added size and missing loc
Refactored the ChatLinks modules so they each have their own files, so they are easier to maintain.
Simplified the creation of action buttons a lot. All the communication back and forth is now automatically handled by the button itself when created, we only have to indicate command IDs and what to do when receiving data.
The size of the link content is sent to be displayed in the tooltip, so the receiver knows how huge the thing will be.
Added missing localization.