[modules/register] Add copyright line to MSP module
In my country, only the copyright symbol itself is legally meaningful in any way and copyright terms are based on my future date of death, rendering a date meaningless for my contributions.

The Berne Convention means these conditions apply to my contributions worldwide, but others' own countries' laws apply to their contributions worldwide. It's a bit of a mess, really.