[modules/register] Use a pre-defined supported list of MSP fields
This list is built from TT and REQUEST_TAB. This helps ensure that any fields the subject has unset in an MSP addon will be properly unset in TRP3 too. The version numbers need to be used to check if the data is empty because the field is empty or empty because TRP3 doesn't reload LibMSP data.