[modules/register] Update MSP glance parsing matches
These look scary, but are more efficient and fix some edge cases of partial glances.

%f (the frontier pattern) is undocumented in Lua 5.1, but was officially documented in 5.2, so it's not going anywhere and is safe to use here.

The first part of the pattern %f[^\n%z] is essentially equivalent to ^ in regexps, matching the start of any line (including the first -- assuming CRLF or LF line endings). %f[\n%z] is essentially equivalent to $ in regexps, matching the end of any line (including the last).