[modules/register] Take slightly more care with MSP PX removal
This will no longer potentially drop prefixes that match the beginnings of names accidentally.

In the MSP rewrites, I intentionally set PX to be usable one of two ways: either removed from the NA field prior to display (TRP sends this form) *or* as an independent field that can be prefixed to NA when desired (XRP sends this form in 2.0.0-beta2+). Both work correctly with this parsing method, but this makes the latter form more reliable if the prefix and the proper beginning of the NA field coincidentally match (i.e., PX="Dr", NA="Dresden").

This is an unlikely corner case, but it's better to fix now, especially since the fix is cheap.