Export psycho traits via MSP "PS" field
These are specified in a machine-readable manner in case an addon wants
to display them like we do.

A "PS" field is used containing a set of traits wrapped in square brackets
with key=value pairs inside.

Example of a built-in trait:
    [trait id="3" value="0.75"]

Example of a custom trait:
    [trait value="0.4" left-name="My Left Trait"
    right-name="My Right Trait"]

The custom trait supports exporting the name, icon, and color data.

The current implementation supports serialising and (largely
untested) deserialising the data to/from this format.

One unresolved challenge is establishing an agreed upon set of IDs for
built-in traits.