Centralise request fields
Also known as "WHY ARE SOME OF THE FIELDS MISSING", turns out that
we're idiots when it comes to requesting things in a sensible manner.

We've mostly been lucky because the inconsistencies in what we get back
are covered by tooltip fields, that and the "worse" list has been the
slash command.