Correctly clear msp.char on registry deletion
deleteProfile was being called in such a way that it would never
emit an event with the mspOwners table populated, thus the handler
in the MSP module would never clear msp.chars for any deleted profiles.

This would mean that profiles which are culled then re-received in the
same UI session would be horribly desynced; the MSP module would see
the data in msp.chars and think it was up to date but the TRP data
would have very few fields present.

I imagine there's a solid reason behind _why_ deletion doesn't
propagate events, including but not limited to the potential of
an infinite loop, so instead we've decided to just delete the
msp.chars entries directly from within deleteProfile. The table
of mspOwners will still be emitted if it can, but the data won't
be present in msp.chars at the time of emission. This shouldn't
be a concern because event handler order is (hopefully) unspecified.