Send formatting tags via MSP DE field
Allows sending of formatting tags like {p} and {h1} via MSP, which
enables MRP to support them in its upcoming browser rework.

Other addons will need to strip them on receipt; we'll have a
library to centralize this eventually, but for now this delivers
the requirement. Other addons represent an extremely minimal
segment of the userbase anyway (if they even work...).