Improve module loading for debug builds
In debug builds, if a module fails to load then the error holds up
the entire lifecycle process of the addon. This is desirable, but
you can indeed have your cake *and* eat it by using xpcall to
forward the error information to the global error handler and *not*
stopping the lifecycle from proceeding.

If the lifecycle doesn't proceed, the addon ends up in a zombie
state where a bunch of stuff doesn't work as expected. This is
a bit annoying for development because, for example, the slash
commands for the UI won't be installed.

Forwarding the errors to the default error handler via xpcall and
CallErrorHandler flags them for review in your error handling
addon (eg. BugSack). They still make loud noises, they're still
very noticeable, but they don't break the addon entirely.