Speculative fix for map anchoring errors
Explicitly initialize the frame as parented to UIParent and remove
its initial anchors.

We're clearing these on module init anyway, so we might as well
see if this approach works. Personally I couldn't reproduce the
issue *without* the ClearAllPoints line removed either, so there
seems to be some other issues at play.