Fix recursive dependency in TOC
The main addon depends upon the registry data, which itself depends
upon the main addon. The fact this works is pure luck, but it's by
no means likely meant to be supported.

As the registry data addon does nothing, we'll instead remove its
dependency upon the main addon. Additionally, flag it as LoD since
there's no need to load the registry data if the main addon is itself
disabled.

Potentially users could now disable the registry data; if that's
actually an issue (I'd *expect* the registry to be optional...)
then we could force-enable it prior to loading from the main
addon perhaps.