Fix toolbar overlaying frames
"Fullscreen" frames like the world map when maximized don't set
the frame strata to FULLSCREEN, which I guess is what used to
happen pre-world map rework. Instead they just bump up their
frame levels, which is why opening things like the talent panel
will show it above the world map.

The best fix for this is to set the strata of the toolbar down to
MEDIUM, which will cause it to no longer overlay such panels. This
also fixes the issue whereby we'd overlay our own TRP main window
with the toolbar if maximized.