Initial commit of name plate module
Implements basic customizations to player name plates. At the
moment this is limited to name text and colouring.

Profiles will be queried when the name plate is initialized for
an ingame unit, and will be updated accordingly when new profile
data is received.

At the moment there's no options in place for changing the behaviour,
nor is there support for titles/icons.