Update name plates on configuration change
Rather than carry on with the old approach of config changes,
I've added an event that's invoked on all config changes. This
means that reacting to generic config changes doesn't require
copy/pasting string keys around the addon all the time.

This change in particular enables me to correctly refresh the
frame on changes to color contrast, but it'll also give me an
easy hook for adding other settings for this module.