Add detection of IC/OOC toggle
This uses a low resolution timer as there's no event for this
change occuring. The timer is set to tick every second, which
is fast enough for the toggle to feel like it works but also
not actually consume unholy amounts of resources.

Did some minor refactoring to handle the resetting of names as
part of this change, since we need to reset them when toggling
this state to ensure those customizations are discarded.