Ability to disable nameplates on addon conflict
There's a blacklist table (CONFLICTING_ADDONS) that will prevent
the module from being enabled if any named addons contained within
are enabled on the users' current addon list. At the moment this
table is empty, but we can add to it if conflicts do arise.

Configuration key registration was moved to initialization and
takes place *before* this conflict test, so that config keys can
be used elsewhere even if we elect to turn ourselves off due to
a conflict. Of course this doesn't help in the case that the
module is disabled.