Add CanCustomizeUnitFrame function
This is intended to be a place where we can place checks to know if
a frame is a valid target for customization. At the moment this just
checks if it's forbidden or not, but if there's any other logic that
needs adding it gives us a one-stop area to place it in.

This new function doesn't check if customizations should or should
not occur due to settings; it's purely to know if we'd explode should
the addon touch the frame in any way or not.