Refactor nameplate module
Split the display handling code to a separate file with a new
base mixin (DisplayProviderMixin). This means adding new display
providers (such as for Kui) can use the same core codebase and
only worry about how it presents the data, rather than how it
gets notified of important changes.

The AddOn_TotalRP3.NamePlates namespace has been exported and houses
common functions for obtaining the values to display, the state of
configuration settings, and functions to trigger explicit updates
or requests.