Refactor property querying code
Add in a separate Display.lua script which for now handles the
backend of obtaining data for unit tokens, and checking if
customizations can be applied to units.

The DecoratorBaseMixin provides higher level methods which wrap
these and will alter their returns to give a consistent look
across decorators, such as a common title cropping length, or
fully qualifying icon paths.