Add support for Kui nameplates
Reorganized the decorators into a subfolder and moved out the
title color constant while I was at it.

Kui nameplates fully integrate with the available settings and
seem to work just fine.

One change still to be resolved (maybe?) is the hide non-RP
plates setting which I've still not implemented.