Experimental improvements to Kui customizations
I noticed during testing that custom colors would sometimes reset
for players that would go in/out of combat in name-only mode while
not having the option enabled to color their name as their
health percentage.

This should fix that, but the approach is quite heavy handed and
might not be desirable. This isn't done by the original module
so it might be worth checking the behaviour there.