Cleanups for Kui integration
After some thought I've determined this hook approach is the
best option for implementing it; the main issue otherwise is that
to work around the title option issue we're left in an arms race
against the core layout since a lot of messages/events it listens
to can update that area of the frame, and these could change at
any time.

The hooks themselves are quite small and, while maybe not the nicest
approach, do have very good and stable results so far.

As our internals (and those of the original separate module) depend
on the "Kui Nameplates: Core" layout I've added a check for this
on module initialization and will display a message if customizations
will be disabled due to this fact.