Prompt to enable guild text when enabling titles
This only takes place if the setting is toggled in the UI, but
this should be enough. Showing a message at startup for setting
that's enabled by default might be spammy, and is hampered by the
fact that KNP settings aren't loaded until *after* we set up
our own stuff.