Protect against forbidden frames
In theory these should never enter the system now, the only vector
is through unit frame name changes. I'd initially thought you could
at least get the parent of a forbidden frame and it would have not
passed the nameplate test later, but that's not allowed so we'll
explicitly check there as the first thing instead.