Remove handling of some NAME_PLATE_* events
No longer forward added/removed/created events to the active
decorator, as the only one that looks like it'll be using it
is the Blizzard one.

Instead, the Blizzard decorator will hook the NamePlateDriverFrame
to determine when to apply modifications. This also gives us another
advantage in that event order isn't deterministic anyway, so it
could have been possible we'd apply updates *before* the base
driver did things.

If the user has a custom nameplate addon installed, the convention
seems to be that they'll unregister events (or themselves hook)
these same functions on the base nameplate driver. This should
remove a bunch of pointless no-ops in those cases regardless.