Don't request profiles if customizations disabled
We'd previously check the state of the global enable/disable
toggle, but not the combined state that indicates if customizations
actually *are* disabled (eg. due to being OOC).

This does cause one minor issue in that if you enable customizations,
requests won't be issued for plates that are already visible.