Implement glances on nameplates, and name cropping
Glances are displayed above nameplate in combat (gapless) form
as small circular buttons, which may be disabled in the settings.

Glances are only shown for actively-targetted units, or ones
that you have very recently mouseovered.

Hovering over the button will show a tooltip with the contents
of the glance, as well options for left/right clicking.

Left clicking will open the glance in a small popup display
which can be moved around and read more easily. This is to
work around the fact that reading a tooltip on a moving player
is actually exceptionally difficult.

Right clicking will open the player's profile. Seemed like a nice
touch.

At the moment, glances are only implemented for Kui nameplates.

Name cropping has been implemented for decorators, capping at the
same length as the title length (which was reduced to 45 characters).

Some code cleanups have occurred to upvalue frequently used
constants.