Add roleplay language field
This field displays an icon in the tooltip for a user indicating
their preferred roleplay language. On an upgrade, this will initialize
to the addon locale if set or will fall back to the client locale.

The setting can be changed from the dashboard as a new dropdown
with the RP status/RP XP options.

The icon will only display in tooltips if it differs from the
users currently selected locale; so you won't see enUS flags if
you're running TRP with the enUS locale selected since the info
isn't all that useful in that case.