Add Luacheck global generation script
Updates the contents of the WoW-provided globals with data
extracted from the game by a third party. This covers 99% of the
API it seems, and matches our existing approach with the giant
luacheck file.

Manual additions were made to cover any other missing global
defines that we hadn't yet sorted out.