Initial conversion to 9.x Backdrop format
This adds a new TRP3_BackdropTemplate frame template and associated
mixin (TRP3_BackdropTemplateMixin) which, for 9.x clients will
transparently mixin the Blizzard-provided BackdropMixin functions
onto a given frame, and for older clients use a custom polyfill
which mirrors the new API provided in 9.x.

Where possible the polyfill stays as close to the 9.x implementation
as possible, but for things that can't be implemented a few no-op
stub functions are provided.

All usages of <Backdrop> XML elements within TRP3 have been changed
to now inherit the appropriate template and specify the expected
KeyValue elements to configure their backdrops. All the backdrop
info tables have been placed in the same file as the new backdrop
implementation, with a naming scheme matching Blizzard's to avoid
unnecessary bikeshedding.

Not fully checked, but with luck all usages of SetBackdrop in
the code (of which there's not many) should automatically just
work so long as they're targetting frames that previously had
backdrops assigned via XML.

Note that a few of our dependencies use SetBackdrop (or <Backdrop>
XML elements); these haven't been updated and so will probably
still break things.

The changes here are also wholly untested and might break everything.