Increase load priority of widgets.xml
This resolves the (final?) issue with the backdrop changes where the
ResizeShadowFrame required the new template, but was instantiated
earlier in the load process.

The widgets.xml file is now loaded first, which is safe since it only
declares templates and mixins and has no dependencies on other areas
of the addon.