Improve handling of improper case in /trp3 open
Before, we'd only attempt to uppercase the first letter of the name and
wouldn't try to case-correct the realm. Similarly upon receiving new
register data we would check if the sender (which will be case-correct)
against a possibly-incorrect user input, and if these didn't match the
UI wouldn't open.

The new logic fixes the above issues by attempting to case-correct both
the requested player name and their realm separately. If the user still
manages to screw this up (by typing the whole thing in caps), then when
register data is updated we'll now do a case-insensitive comparison
against the sender and open the UI based on that identifier rather than
the user input.