Put .luacheckrc on a diet
Rather than have that huge monstrosity of a generated file (that we
don't regenerate anyway, wink wink) this slims it down to the precise
globals that we're actually using.

Additionally, this changes how some of our write globals are defined so
we can better document the things that we modify publicly - eg. instead
of saying that "WIM" is writeable we now scope it to just the few
functions that we hook.