Migrate from Travis to GitHub Actions (#440)
* Migrate from Travis to GitHub Actions

* Update luacheck action to use upstream repo

Needs a new tag upstream to support annotations, so for now we'll pin it
to a commit and go back later.

* Put .luacheckrc on a diet

Rather than have that huge monstrosity of a generated file (that we
don't regenerate anyway, wink wink) this slims it down to the precise
globals that we're actually using.

Additionally, this changes how some of our write globals are defined so
we can better document the things that we modify publicly - eg. instead
of saying that "WIM" is writeable we now scope it to just the few
functions that we hook.

* Remove the test/localizator from excludes

These died ages ago. Thanks Elly <3

* Split editorconfig into its own composite action

This also adds annotation support thanks to the power of awk, so we can
now get source locations in PRs when someone mis-indents something.

* Make patterns more strict