Prevent double-receipt of profile comms via MSP
This fixes the initial double-request of profiles via both the TRP
and MSP protocols when mousing over units or opening them in via the
/trp3 open command.

Before, we'd send a request out to the other player via both protocols
at once. The MSP layer has logic that causes it to ignore responses
received by people running TRP, however that doesn't change the fact
that we've still asked them for their entire profile - so we end up
receiving (in the worst case) twice as much data as actually needed.

To fix this a small change is made so that all MSP requests go through
a single, previously unused, function. When a request is sent via this
path it first checks if we've received the players' VA field and, if
not, will first issue a request _solely_ for that field.

Upon receiving a response we then consult the table to know if this was
a "hello" request and, if so, re-request the full data from the other
player if they're not running TRP.

This adds a bit of latency to initial MSP comms however in brief testing
yesterday this was unnoticeable.