Reformat full title logic
Allow empty strings and nil returns from the profile data and
UnitPVPName, only filtering them once when we reach the actual branch
for adding the line to the tooltip.

This makes the intent of the code a bit clearer by ensuring that we only
show the title if it both exists, and if we've not received an empty
string.

This is also safe against any weird behaviours in UnitPVPName that _may_
cause it to return empty strings; though this is entirely hypothetical.