Allow PR branch check step to fail
PRs that are submitted from forked repositories run the actions within
the fork, instead of our repository, and so don't have permissions to
access the resources required to post comments within the PR, or so it
seems.

As we're mostly using this step as an internal warning only, we can
allow it to fail and just apply due diligence for now when checking
external PRs.