Slim down the issue template (#481)
This removes a bit of the wordy cruft that may have been impacting
peoples ability to correctly fill in - or leading people towards outright
deleting the contents of - the issue template. While we lose some of the
verbosity in the instructions, this should be clear enough at a quick
glance and requires a lot less reading for users to submit an issue, and
if they're not sure then we can just provide those instructions as
needed (or move them to the wiki).