Add UNIT_TYPE enumeration
This replaces the constants previously used for TRP3_API.ui.misc.TYPE_*,
all usages in the addon have been changed to use the enum.

The values present in the enum are numeric and so differ from the old
string values like "PET" or "NPC" - hopefully nothing depends on the
precise values, but if so we could change them back.