Improve handling of pet units to getTargetType
The original logic would return the PET type for a unit if the queried
unit token was either the players' pet, or that of another player. This
meant that Hunters with the Animal Companion would never return the
"PET" type if this function were used to query their second pet, as it
doesn't have an associated unit token.

The new logic now simply checks two things - firstly, if the unit is
controlled by a player, and secondly if the unit has a "Pet" type GUID.
This works with pets summoned through the Animal Companion talent and
presumably in other cases where a player can have multiple active pets,
and shouldn't falsely trigger for companions of NPCs.