Add getCurrentSecondaryPetQueryLine
This mirrors the behaviour of getCurrentPetQueryLine except is intended
for use with "secondary" pets like those summoned through the Animal
Companion talent. Currently this is limited to only working with said
talent, and will return the query line associated with whatever profile
has been associated with the pet in the first Hunter stable slot.