Add support for primary/secondary pets to exchange
The functions and constants for dealing with the pet unit were renamed
to refer to the "primary" pet, whereas "secondary" pet is now that of a
second companion if applicable.

This surprisingly actually just works as-is and implements the ability
for players to view companion profiles associated with other players
secondary (Animal Companion) pets.