Refactor parts of the pet browser
The functions that were present on the frame mixin for dealing with the
stable master APIs have been moved into local functions as they didn't
require any state present on the frame, and didn't need to be publicly
accessible.

A new AddOn_TotalRP3.Ui table was added with a few functions for
retrieving the pet browser frame and checking if it should be enabled
under current player conditions; on Classic in particular the frame is
now disabled and won't be created, which should resolve a significant
number of issues there.