Improve pet browser warnings
The warning for pets that are bound to an existing profile has been
improved for clarity and to say what'll happen if the user assigns a
profile to the pet. An additional warning was added to deal with the
case where the selected pet appears to have not been renamed - though
it's likely that this isn't perfect since some tames probably don't
share the same name as their family, it should at least filter out the
common case.