Apply level squish to as-yet-unsummoned pets
Pet levels are squished to 9.0 only after they're summoned for the first
time, and until that occurs the stable master APIs will return
pre-squish levels.

To avoid any unnecessary tickets about "this says my pet is level 110"
we'll forcefully squish levels returned from the API if they're greater
than the maximum attainable player level for the latest released
expansion.

Naturally there's an edge case if the player had a level 60 pet
pre-squish - as that'll show as 60 in the UI - but that should be much
rarer than the case of someone looking at a pet in the UI that's been in
their stables for the past few years and seeing a wrong level.