Simplify handling of pet browser tooltips
Instead of having the OnEnter/OnLeave scripts be responsible for
managing the tooltip frame, we now leave that responsibility to
UpdateTooltipVisualization.

To activate the tooltip for a button a new SetTooltipSlot method is
provided which will, if given a valid slot index, show the tooltip
anchored to the button for the specified slot.

Additionally, GameTooltip is no longer used to avoid potential taint
issues.