UI tweaks for pets bound to a profile
Pets bound to a profile will now have an overlay icon of a lock to make
it a bit clearer that there's something special about them, in case a
user doesn't read the tooltip.

The lock icon was chosen because it's a pain to find anything else
that's even remotely relevant for this type of indicator.