Add Pronouns preset to Additional Information
This implements full support for the pronoun field as a new preset in
the additional information section and has been tested with basic TRP
and MSP comms.

The preset is mapped to the new "PN" field in MSP, which was suggested a
while ago by Etarna.

Included is also a small set of bugfixes for invalid icons being used in
Classic when receiving other info fields such as the house name.