List battle pets irrespective of journal filters
The C_PetJournal API is a stateful API with regards to applied search
filters and sorting options. This leads to a small bug in our browser
where if the user has filtered their journal, we may incorrectly not
show pets that are actually valid for companion assignment.

This (lengthy) PR resolves this by resetting all filter state on the
pet journal while we collect all the players' battle pets, and then
restores it. It does so as safely as possible so that if any errors
occur we restore as much of the original state as possible.

An additional change is to make the companion browsers no longer use
safeMatch and instead use plain text matches, this solves another
issue where invalid patterns would have weird effects on the search
results for the browsers. Odds are users don't know about nor care about
pattern functionality for searches.