Check for English misc field names in MSP/Tooltips
This resolves potential edge cases where localizations change for any
MI fields like nicknames or pronouns. If we launch an update with no
localization for the "Pronouns" info preset, people will embed that into
their profiles and - if a later update then changes the string - the
original could for non-english locales no longer correctly transmit
the original Pronouns field info.

This is a bit of a quick hack. Ideally we'd store the preset ID in the
MI stuff, but for now we'll just check for both english and localized
strings.