Fix initialization of default profile LC field
The LC field was reading the TRP3_Configuration table in the main body
of the script before it's loaded - this works silently because the
config code (unhelpfully) defaults the table to an empty one, leading
to slient bugs such as this.

LC initialization now defaults to client locale initially and is then
updated once WORKFLOW_ON_LOADED fires. We don't increment the vernum
of the default profile and assume this is called early enough that no
one will have possibly noticed the LC field changing as a result.