Secondary Pet Companions (#482)
* Add UNIT_TYPE enumeration

This replaces the constants previously used for TRP3_API.ui.misc.TYPE_*,
all usages in the addon have been changed to use the enum.

The values present in the enum are numeric and so differ from the old
string values like "PET" or "NPC" - hopefully nothing depends on the
precise values, but if so we could change them back.

* Improve handling of pet units to getTargetType

The original logic would return the PET type for a unit if the queried
unit token was either the players' pet, or that of another player. This
meant that Hunters with the Animal Companion would never return the
"PET" type if this function were used to query their second pet, as it
doesn't have an associated unit token.

The new logic now simply checks two things - firstly, if the unit is
controlled by a player, and secondly if the unit has a "Pet" type GUID.
This works with pets summoned through the Animal Companion talent and
presumably in other cases where a player can have multiple active pets,
and shouldn't falsely trigger for companions of NPCs.

* Add getCurrentSecondaryPetQueryLine

This mirrors the behaviour of getCurrentPetQueryLine except is intended
for use with "secondary" pets like those summoned through the Animal
Companion talent. Currently this is limited to only working with said
talent, and will return the query line associated with whatever profile
has been associated with the pet in the first Hunter stable slot.

* Add support for primary/secondary pets to exchange

The functions and constants for dealing with the pet unit were renamed
to refer to the "primary" pet, whereas "secondary" pet is now that of a
second companion if applicable.

This surprisingly actually just works as-is and implements the ability
for players to view companion profiles associated with other players
secondary (Animal Companion) pets.

* Fix unused locals

* Restore values to UNIT_TYPE enum

We do actually store the constants associated with them in some places,
so we can't convert these to numbers and need to use their original
strings.

* Add companion browser for Hunter pets

This provides a way of binding companion profiles to hunter pets based
upon the information retrieved from the Stable APIs without users
needing to manually summon and target the pet they wish to bind to.

* Indent the giant localization warning block

* Fix license indentation

* Refactor parts of the pet browser

The functions that were present on the frame mixin for dealing with the
stable master APIs have been moved into local functions as they didn't
require any state present on the frame, and didn't need to be publicly
accessible.

A new AddOn_TotalRP3.Ui table was added with a few functions for
retrieving the pet browser frame and checking if it should be enabled
under current player conditions; on Classic in particular the frame is
now disabled and won't be created, which should resolve a significant
number of issues there.

* Improve pet browser warnings

The warning for pets that are bound to an existing profile has been
improved for clarity and to say what'll happen if the user assigns a
profile to the pet. An additional warning was added to deal with the
case where the selected pet appears to have not been renamed - though
it's likely that this isn't perfect since some tames probably don't
share the same name as their family, it should at least filter out the
common case.

* Apply level squish to as-yet-unsummoned pets

Pet levels are squished to 9.0 only after they're summoned for the first
time, and until that occurs the stable master APIs will return
pre-squish levels.

To avoid any unnecessary tickets about "this says my pet is level 110"
we'll forcefully squish levels returned from the API if they're greater
than the maximum attainable player level for the latest released
expansion.

Naturally there's an edge case if the player had a level 60 pet
pre-squish - as that'll show as 60 in the UI - but that should be much
rarer than the case of someone looking at a pet in the UI that's been in
their stables for the past few years and seeing a wrong level.

* Simplify handling of pet browser tooltips

Instead of having the OnEnter/OnLeave scripts be responsible for
managing the tooltip frame, we now leave that responsibility to
UpdateTooltipVisualization.

To activate the tooltip for a button a new SetTooltipSlot method is
provided which will, if given a valid slot index, show the tooltip
anchored to the button for the specified slot.

Additionally, GameTooltip is no longer used to avoid potential taint
issues.

* Clean up usage of GameTooltip in Dashboard

This is just as a taint-preventative measure. Didn't really want to
sneak it in with these changes, but it's small enough.

* Improve handling of dialog callbacks

Instead of having three separate callbacks for accept/cancel/close
(where the last is always invoked), the pet browser frame now just has
one callback which exposes the result as an enumeration (Accept/Close)
and provides additional data in the accept case.

* Minor UI tweaks

Made the "Assign Profile" button just say "Assign" and shrunk it down
accordingly, as well as updated the OnClick handlers for the last set of
dialog changes.

* Minor style tweaks to browser icons

* Pass current profile ID through to browser dialog

This will be used for some filtering logic to improve the behaviour of
the list.

* Filter pets assigned to the current profile

This attempts to simplify and reduce the number of icons a user has to
browse through when selecting a pet by hiding those that are already
assigned to the profile selected by the user.

* UI tweaks for pets bound to a profile

Pets bound to a profile will now have an overlay icon of a lock to make
it a bit clearer that there's something special about them, in case a
user doesn't read the tooltip.

The lock icon was chosen because it's a pain to find anything else
that's even remotely relevant for this type of indicator.

* Call SetTooltipIcon directly from handlers

* Hide overlay icon when decorating empty slots

* Revert getCurrentPetQueryLine renaming

Just in case for backward compat reasons. Having yet another alias for
backwards compatibility that never gets cleared up isn't appealing.

* Cleanup UNIT_TYPE.MOUNT todo

Co-authored-by: Solanya <17127080+Solanya@users.noreply.github.com>